refactor(runner-core): split kernel tests
